Add robust load_config, load_data, preprocess_data and tests

+++ src/train/train_model.py
@@ -1,23 +1,24 @@
-import pandas as pd
 import mlflow
 import mlflow.sklearn
+import pandas as pd
 from sklearn.ensemble import RandomForestClassifier
-from sklearn.model_selection import train_test_split
 from sklearn.metrics import accuracy_score
-import os
+from sklearn.model_selection import train_test_split
 
 mlflow.set_experiment("churn-experiment")
 
 df = pd.read_csv("data/raw/telco_churn.csv")
-df = df[df['TotalCharges'] != ' ']
-df['TotalCharges'] = df['TotalCharges'].astype(float)
-df['Churn'] = df['Churn'].map({'Yes': 1, 'No': 0})
+df = df[df["TotalCharges"] != " "]
+df["TotalCharges"] = df["TotalCharges"].astype(float)
+df["Churn"] = df["Churn"].map({"Yes": 1, "No": 0})
 df = pd.get_dummies(df)
 
-X = df.drop('Churn', axis=1)
-y = df['Churn']
+X = df.drop("Churn", axis=1)
+y = df["Churn"]
 
-X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
+X_train, X_test, y_train, y_test = train_test_split(
+    X, y, test_size=0.2, random_state=42
+)
 
 with mlflow.start_run():
     clf = RandomForestClassifier()
